remove: register activity
