+++ app/src/main/java/ro/code4/deurgenta/ui/login/LoginActivity.kt
@@ -1,12 +1,18 @@
 package ro.code4.deurgenta.ui.login
 
 import android.os.Bundle
+import androidx.fragment.app.commit
+import androidx.fragment.app.replace
 import kotlinx.android.synthetic.main.activity_login.*
+import kotlinx.android.synthetic.main.fragment_register.*
 import org.koin.android.viewmodel.ext.android.viewModel
 import ro.code4.deurgenta.R
 import ro.code4.deurgenta.helper.replaceFragment
 import ro.code4.deurgenta.helper.startActivityWithoutTrace
 import ro.code4.deurgenta.ui.base.BaseAnalyticsActivity
+import ro.code4.deurgenta.ui.register.RegisterCompletedFragment
+import ro.code4.deurgenta.ui.register.RegisterFragment
+import ro.code4.deurgenta.ui.register.RegisterViewModel
 
 class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
 
@@ -15,6 +21,7 @@ class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
     override val screenName: Int
         get() = R.string.analytics_title_login
 
+    private val registerViewModel: RegisterViewModel by viewModel()
     override val viewModel: LoginViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -22,6 +29,7 @@ class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
 
         clickListenersSetup()
         loginUserObservable()
+        registerObservables()
     }
 
     private fun clickListenersSetup() {
@@ -29,7 +37,7 @@ class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
             showLoginFormFragment()
         }
         signupButton.setOnClickListener {
-            viewModel.signup()
+            showRegistrationFormFragment()
         }
     }
 
@@ -37,6 +45,20 @@ class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
         supportFragmentManager.replaceFragment(
             R.id.login_container,
             LoginFormFragment()
+        )
+    }
+
+    private fun showRegistrationFormFragment() {
+        supportFragmentManager.replaceFragment(
+            R.id.login_container,
+            RegisterFragment()
+        )
+    }
+
+    private fun showRegistrationCompletedFragment() {
+        supportFragmentManager.replaceFragment(
+            R.id.login_container,
+            RegisterCompletedFragment()
         )
     }
 
@@ -50,6 +72,19 @@ class LoginActivity : BaseAnalyticsActivity<LoginViewModel>() {
                     showDefaultErrorSnackBar(loginButton)
 
                     loginButton.isEnabled = true
+                }
+            )
+        })
+    }
+
+    private fun registerObservables() {
+        registerViewModel.registered().observe(this, {
+            it.handle(
+                onSuccess = {
+                    showRegistrationCompletedFragment()
+                },
+                onFailure = { error ->
+                    showDefaultErrorSnackBar(submitButton)
                 }
             )
         })
