+++ app/src/main/java/ro/code4/deurgenta/ui/login/LoginViewModel.kt
@@ -10,7 +10,6 @@ import ro.code4.deurgenta.repositories.Repository
 import ro.code4.deurgenta.ui.base.BaseViewModel
 import ro.code4.deurgenta.ui.main.MainActivity
 import ro.code4.deurgenta.ui.onboarding.OnboardingActivity
-import ro.code4.deurgenta.ui.register.RegisterActivity
 
 class LoginViewModel : BaseViewModel() {
 
@@ -26,11 +25,6 @@ class LoginViewModel : BaseViewModel() {
             true -> MainActivity::class.java
             false -> OnboardingActivity::class.java
         }
-        loginLiveData.postValue(Result.Success(nextActivity))
-    }
-
-    fun signup() {
-        val nextActivity = RegisterActivity::class.java
         loginLiveData.postValue(Result.Success(nextActivity))
     }
