--- app/src/main/res/layout/activity_register.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        tools:context="ro.code4.deurgenta.ui.register.RegisterActivity"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/colorPrimaryLight"
-            android:clipChildren="false"
-            android:clipToPadding="false"
-            android:padding="@dimen/margin">
-
-
-        <ImageView
-                android:id="@+id/closeImageView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:srcCompat="@drawable/ic_close"
-                android:layout_marginTop="16dp" />
-
-        <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/fragment_container_view"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintTop_toBottomOf="@+id/closeImageView"
-                app:layout_constraintStart_toStartOf="parent"
-                android:layout_marginTop="16dp">
-
-
-        </androidx.fragment.app.FragmentContainerView>
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</ScrollView>
