+++ app/src/main/res/layout/fragment_register.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context="ro.code4.deurgenta.ui.register.RegisterActivity"
+        tools:context="ro.code4.deurgenta.ui.login.LoginActivity"
         xmlns:app="http://schemas.android.com/apk/res-auto">
     <data>
         <import type="android.view.View"/>
@@ -13,6 +13,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@color/colorPrimaryLight"
+            android:padding="@dimen/margin"
             android:clipChildren="false"
             android:clipToPadding="false">
